Keep the fresh_db status tests in their own module (#1718)
A module that mixes the module-scoped db fixture with the function-scoped
fresh_db one can hang, because fresh_db drops all tables while the module-scoped
connection is still open. Move the three new tests to a sibling module, following
the *_fresh_db.py convention the repo already uses.

One existing case asserted the behaviour this branch fixes: at a moment when only
the sensor's most recent data is not knowable yet, it expected "no data recorded",
where the sensor does hold older data we could know about by then. It now expects
that data, 12 hours old.


# Conflicts:
#	flexmeasures/api/common/schemas/tests/test_sensor_data_schema.py
